Clarity for Appendix B1 (#145)
* Clarity for Appendix B1

I found myself struggling with the language in B1, so I reworked the diagrams to label the AS and RS in their respective domains.  The language has been carried into the text, as well.

If this is helpful, I'll make the same changes for B2 and other places in the text where more clarity is required.

* Update draft-ietf-oauth-identity-chaining.md

---------

Co-authored-by: Arndt Schwenkschuster <17650715+arndt-s@users.noreply.github.com>